scripts/bench: add tag option -t
Previously, the benchmark's name consisted of the implementation
identifier according to UIOP. Now, when a tag is provided, the name
consists of the implementation identifier followed by a hyphen and the
tag.

If the tag contains any double-quote characters, the Lisp implementation
will probably fail to run.

* scripts/bench: Parse -t option, setting tag variable appropriately.
(tag): New variable.
(run): Set benchmark name according to tag variable.