quaviver, schubfach, dragonbox: conform SIGN types
Previously, SIGN was normally declared as (INTEGER -1 1), but there is
no need to be unduly restrictive; now it may be a FIXNUM.
Also, the type of SIGN in %INTEGER-ENCODE-FLOAT was previously
accidentally declared as (OR FIXNUM KEYWORD); now it is fixed.

* code/float-integer-2.lisp (%integer-decode-float):
* code/schubfach/implementation.lisp (%schubfach):
* code/dragonbox/implementation.lisp (%nearest, %directed): Declare SIGN
as FIXNUM rather than (INTEGER -1 1).
* code/integer-float-2.lisp (%integer-encode-float): Declare SIGN as
FIXNUM rather than the accidental (OR FIXNUM KEYWORD).